+++ app/Http/Controllers/FileDownloadController.php
@@ -48,13 +48,15 @@ class FileDownloadController extends Controller
 
         $ascii_name = Str::ascii($file->name); // ascii name is required for the response as it is mandatory for the Symfony binary response
 
+        $headers = [
+            'Content-Type' => $file->mime_type
+        ];
+
         if (strtolower($request->method()) === 'head') {
-            return response('', 200, ['Content-Length' => 0, 'Content-Type' => $file->mime_type ]);
+            return response()->head($headers);
         }
 
-        return response()->download($file->absolute_path, $ascii_name, [
-            'Content-Type' => $file->mime_type
-        ]);
+        return response()->download($file->absolute_path, $ascii_name, $headers);
     }
 
     /**
